moved Types to a separate file
