+++ SourceryRuntime/Sources/Types_Linux.swift
@@ -1,0 +1,209 @@
+//
+// Created by Krzysztof Zablocki on 31/12/2016.
+// Copyright (c) 2016 Pixle. All rights reserved.
+//
+#if !canImport(ObjectiveC)
+import Foundation
+// For DynamicMemberLookup we need to import Stencil,
+// however, this is different from SourceryRuntime.content.generated.swift, because
+// it cannot reference Stencil
+import Stencil
+
+// sourcery: skipJSExport
+/// Collection of scanned types for accessing in templates
+public final class Types: NSObject, SourceryModel, Diffable, DynamicMemberLookup {
+    public subscript(dynamicMember member: String) -> Any? {
+        switch member {
+            case "types":
+                return types
+            case "enums":
+                return enums
+            case "all":
+                return all
+            case "protocols":
+                return protocols
+            case "classes":
+                return classes
+            case "structs":
+                return structs
+            case "extensions":
+                return extensions
+            case "implementing":
+                return implementing
+            case "inheriting":
+                return inheriting
+            case "based":
+                return based
+            default:
+                fatalError("unable to lookup: \(member) in \(self)")
+        }
+    }
+
+    /// :nodoc:
+    public let types: [Type]
+
+    /// All known typealiases
+    public let typealiases: [Typealias]
+
+    /// :nodoc:
+    public init(types: [Type], typealiases: [Typealias] = []) {
+        self.types = types
+        self.typealiases = typealiases
+    }
+
+    /// :nodoc:
+    override public var description: String {
+        var string = "\(Swift.type(of: self)): "
+        string += "types = \(String(describing: self.types)), "
+        string += "typealiases = \(String(describing: self.typealiases))"
+        return string
+    }
+
+    public func diffAgainst(_ object: Any?) -> DiffableResult {
+        let results = DiffableResult()
+        guard let castObject = object as? Types else {
+            results.append("Incorrect type <expected: Types, received: \(Swift.type(of: object))>")
+            return results
+        }
+        results.append(contentsOf: DiffableResult(identifier: "types").trackDifference(actual: self.types, expected: castObject.types))
+        results.append(contentsOf: DiffableResult(identifier: "typealiases").trackDifference(actual: self.typealiases, expected: castObject.typealiases))
+        return results
+    }
+
+    public override var hash: Int {
+        var hasher = Hasher()
+        hasher.combine(self.types)
+        hasher.combine(self.typealiases)
+        return hasher.finalize()
+    }
+
+    /// :nodoc:
+    public override func isEqual(_ object: Any?) -> Bool {
+        guard let rhs = object as? Types else { return false }
+        if self.types != rhs.types { return false }
+        if self.typealiases != rhs.typealiases { return false }
+        return true
+    }
+
+// sourcery:inline:Types.AutoCoding
+
+        /// :nodoc:
+        required public init?(coder aDecoder: NSCoder) {
+            guard let types: [Type] = aDecoder.decode(forKey: "types") else { 
+                withVaList(["types"]) { arguments in
+                    NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: arguments)
+                }
+                fatalError()
+             }; self.types = types
+            guard let typealiases: [Typealias] = aDecoder.decode(forKey: "typealiases") else { 
+                withVaList(["typealiases"]) { arguments in
+                    NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: arguments)
+                }
+                fatalError()
+             }; self.typealiases = typealiases
+        }
+
+        /// :nodoc:
+        public func encode(with aCoder: NSCoder) {
+            aCoder.encode(self.types, forKey: "types")
+            aCoder.encode(self.typealiases, forKey: "typealiases")
+        }
+// sourcery:end
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// :nodoc:
+    public lazy internal(set) var typesByName: [String: Type] = {
+        var typesByName = [String: Type]()
+        self.types.forEach { typesByName[$0.globalName] = $0 }
+        return typesByName
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// :nodoc:
+    public lazy internal(set) var typesaliasesByName: [String: Typealias] = {
+        var typesaliasesByName = [String: Typealias]()
+        self.typealiases.forEach { typesaliasesByName[$0.name] = $0 }
+        return typesaliasesByName
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known types, excluding protocols or protocol compositions.
+    public lazy internal(set) var all: [Type] = {
+        return self.types.filter { !($0 is Protocol || $0 is ProtocolComposition) }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known protocols
+    public lazy internal(set) var protocols: [Protocol] = {
+        return self.types.compactMap { $0 as? Protocol }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known protocol compositions
+    public lazy internal(set) var protocolCompositions: [ProtocolComposition] = {
+        return self.types.compactMap { $0 as? ProtocolComposition }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known classes
+    public lazy internal(set) var classes: [Class] = {
+        return self.all.compactMap { $0 as? Class }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known structs
+    public lazy internal(set) var structs: [Struct] = {
+        return self.all.compactMap { $0 as? Struct }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known enums
+    public lazy internal(set) var enums: [Enum] = {
+        return self.all.compactMap { $0 as? Enum }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// All known extensions
+    public lazy internal(set) var extensions: [Type] = {
+        return self.all.compactMap { $0.isExtension ? $0 : nil }
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// Types based on any other type, grouped by its name, even if they are not known.
+    /// `types.based.MyType` returns list of types based on `MyType`
+    public lazy internal(set) var based: TypesCollection = {
+        TypesCollection(
+            types: self.types,
+            collection: { Array($0.based.keys) }
+        )
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// Classes inheriting from any known class, grouped by its name.
+    /// `types.inheriting.MyClass` returns list of types inheriting from `MyClass`
+    public lazy internal(set) var inheriting: TypesCollection = {
+        TypesCollection(
+            types: self.types,
+            collection: { Array($0.inherits.keys) },
+            validate: { type in
+                guard type is Class else {
+                    throw "\(type.name) is not a class and should be used with `implementing` or `based`"
+                }
+            })
+    }()
+
+    // sourcery: skipDescription, skipEquality, skipCoding
+    /// Types implementing known protocol, grouped by its name.
+    /// `types.implementing.MyProtocol` returns list of types implementing `MyProtocol`
+    public lazy internal(set) var implementing: TypesCollection = {
+        TypesCollection(
+            types: self.types,
+            collection: { Array($0.implements.keys) },
+            validate: { type in
+                guard type is Protocol else {
+                    throw "\(type.name) is a class and should be used with `inheriting` or `based`"
+                }
+        })
+    }()
+}
+#endif
